refactor: move tests from test.py to test_decide.py, a more suitable filename, closing #77
